requirements.txt
discord.py
matplotlib
SQLAlchemy
requests
